iNZightRegression: Tools for Exploring Regression Models with 'iNZight'

Provides a suite of functions to use with regression models, including summaries, residual plots, and factor comparisons. Used as part of the Model Fitting module of 'iNZight', a graphical user interface providing easy exploration and visualisation of data for students of statistics, available in both desktop and online versions.

See also

Author

Maintainer: Tom Elliott tom.elliott@auckland.ac.nz (ORCID)

Authors:

Ken Deng

Simon Potter

David Banks

Other contributors:

Danny Chang [contributor]